kakao확정하기 메시지 전송 기능 성공--테스트 코드 X

+++ src/main/java/com/moyeobwayo/moyeobwayo/Repository/KakaoProfileRepository.java
@@ -3,5 +3,5 @@ package com.moyeobwayo.moyeobwayo.Repository;
 import com.moyeobwayo.moyeobwayo.Domain.KakaoProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KakaoProfileRepository extends JpaRepository<KakaoProfile, Long> {
+public interface KakaoProfileRepository extends JpaRepository<KakaoProfile, Integer> {
 }
